protocol: move trait impls into impl_trait mod
