+++ protocol/src/impl_trait.rs
@@ -1,1 +1,2 @@
-mod player_context;
+mod context;
+mod player;
