+++ protocol/src/lib.rs
@@ -3,6 +3,4 @@
 
 mod impl_trait;
 
-mod conversion;
-
 include!(concat!(env!("OUT_DIR"), "/mod.rs"));
